backend/requirements.txt
aiofiles==23.1.0
amqp==5.2.0
anyio==4.3.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
billiard==3.6.4.0
blis==0.7.11
catalogue==2.0.10
celery==5.2.7
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
contourpy==1.2.1
cycler==0.12.1
cymem==2.0.8
DAWG-Python==0.7.2
docopt==0.6.2
executing==2.0.1
fastapi==0.95.0
# fasttext==0.9.2
flower==1.2.0
fonttools==4.51.0
gitdb==4.0.11
GitPython==3.1.43
graphviz==0.20.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
humanize==4.9.0
icecream==2.1.3
idna==3.7
iniconfig==2.0.0
jellyfish==1.0.3
Jinja2==3.1.2
joblib==1.4.0
kiwisolver==1.4.5
kombu==5.3.7
langcodes==3.3.0
MarkupSafe==2.1.5
matplotlib==3.8.4
murmurhash==1.0.10
networkx==3.3
nltk==3.8.1
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pluggy==1.4.0
preshed==3.0.9
prometheus_client==0.20.0
prompt-toolkit==3.0.43
pybind11==2.12.0
pydantic==1.10.15
Pygments==2.17.2
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
pymorphy3==2.0.1
pymorphy3-dicts-ru==2.4.417150.4580142
pyparsing==3.1.2
pytest==7.2.2
pytextrank==3.3.0
python-dateutil==2.9.0.post0
pytz==2024.1
redis==4.5.4
regex==2024.4.16
requests==2.28.2
rfc3986==1.5.0
ru-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/ru_core_news_sm-3.7.0/ru_core_news_sm-3.7.0-py3-none-any.whl
scipy==1.13.0
segtok==1.5.11
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.1
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
starlette==0.26.1
tabulate==0.9.0
thinc==8.2.3
tornado==6.4
tqdm==4.66.2
typer==0.9.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==1.26.18
uvicorn==0.21.1
vine==5.1.0
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
yake==0.4.8
fasttext-wheel
